This attempt to replace qtimer with linux timer
In this version, an attempt is made to replace the qtimer with accurate linux timer (set up by Dr. Bern Porr). The idea is to rewrite Bern's linux timer class in qobject and create a timeout() signal as in qtimer. So that the timmer event can be easily trigger motion of other class through signal (timeout()) to slot