200310-19-00 srvr end based on rs evt instd timer
As we gonna put srvr end on Android, we can't use the qtimer. not even the qprecisetimer. System API timer can't be used as well. The reason is that android is less "real-time" than linux --- the android sys reflash usually every 5ms.
Yet we can make the motion based on receiving the signal. the signal sent from the Raspberry PI is real-timed by the pi