200310-19-30 srvr end faster rsv bsd with ll slot
To make the thread delay smaller with event base (qt signal-slot), the main slot (pong gaming) is spit in several parallel slot triggered by rsv event. The main slot currently only works for move the ball, which is very fast and even executed twice in 1 ms as long as the sig is fired fast enough. all the direction counting, bouncing and scoring are in ll slot